Move boost pad locations to domain module
